+++ packages/react-ui/src/css/messages.css
@@ -208,13 +208,15 @@
 }
 
 .copilotKitImageRenderingImage {
-  max-width: 300px;
-  max-height: 300px;
+  max-width: 80px;
+  max-height: 80px;
   width: auto;
   height: auto;
   object-fit: cover;
   border-radius: 12px;
+  background-color: var(--copilot-kit-input-background-color);
   box-shadow: var(--copilot-kit-shadow-sm);
+  cursor: pointer;
 }
 
 .copilotKitImageRenderingContent {
